+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/decode/internal/DecoderInterceptor.kt
@@ -1,20 +1,4 @@
-/*
- * Copyright (C) 2024 panpf <panpfpanpf@outlook.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.panpf.sketch.request.internal
+package com.github.panpf.sketch.decode.internal
 
 import androidx.annotation.MainThread
 import com.github.panpf.sketch.request.ImageData
@@ -24,9 +8,8 @@ import kotlinx.coroutines.withContext
 /**
  * The final request interceptor is used to actually execute the request
  *
- * @see com.github.panpf.sketch.core.common.test.request.internal.DecoderInterceptorTest
+ * @see com.github.panpf.sketch.core.common.test.decode.internal.DecoderInterceptorTest
  */
-// TODO move to decode package
 class DecoderInterceptor : Interceptor {
 
     companion object Companion {